Fix: moved categories to dao.ts and gate.ts
