--- apps/website/constants/categories.ts
@@ -1,29 +1,0 @@
-export const CATEGORIES = {
-  DAOS: [
-    'Community/Social',
-    'Data',
-    'Financial ',
-    'Governance',
-    'HR',
-    'Infrastructure',
-    'Operating System',
-    'Collaboration',
-    'Services ',
-    'Grants',
-    'Treasury',
-    'Investment',
-    'Media',
-    'Product ',
-    'Protocol',
-    'Research',
-    'Bio',
-  ],
-  GATES: [
-    'Design',
-    'Development',
-    'Business Development',
-    'Marketing',
-    'Education',
-    'Onboarding',
-  ],
-};
