+++ apps/website/constants/dao.ts
@@ -1,0 +1,19 @@
+export const CATEGORIES = [
+  'Community/Social',
+  'Data',
+  'Financial ',
+  'Governance',
+  'HR',
+  'Infrastructure',
+  'Operating System',
+  'Collaboration',
+  'Services ',
+  'Grants',
+  'Treasury',
+  'Investment',
+  'Media',
+  'Product ',
+  'Protocol',
+  'Research',
+  'Bio',
+];
