+++ apps/website/constants/gate.ts
@@ -1,0 +1,8 @@
+export const CATEGORIES = [
+  'Design',
+  'Development',
+  'Business Development',
+  'Marketing',
+  'Education',
+  'Onboarding',
+];
